pyproject.toml
[tool.poetry]
name = "rest"
version = "0.1.0"
description = ""
authors = ["Azat Muzapbar <azamuzapbar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
phonenumbers = "^8.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
